+++ Symphony/render/gl/GLIndexBuffer.cpp
@@ -2,12 +2,12 @@
 
 namespace symphony
 {
-	GLIndexBuffer::GLIndexBuffer(const std::vector<uint16_t>& vertices)
+	GLIndexBuffer::GLIndexBuffer(const std::vector<uint32_t>& vertices)
 		: IndexBuffer(vertices)
 	{
 		glCreateBuffers(1, &m_BufferHandle);
 		glBindBuffer(GL_ARRAY_BUFFER, m_BufferHandle);
-		glBufferData(GL_ARRAY_BUFFER, vertices.size() * sizeof(uint16_t), vertices.data(), GL_STATIC_DRAW);
+		glBufferData(GL_ARRAY_BUFFER, vertices.size() * sizeof(uint32_t), vertices.data(), GL_STATIC_DRAW);
 	}
 
 	GLIndexBuffer::~GLIndexBuffer()
